+++ src/git.rs
@@ -5,13 +5,14 @@ pub fn can_access_remote(
     repo_str: &str,
     inherit_stderr: bool,
     report_git_errors: bool,
+    git_executable: &String,
 ) -> Result<bool, String> {
     let stderr = match inherit_stderr {
         true => Stdio::inherit(),
         false => Stdio::null(),
     };
 
-    let ls_remote_output = Command::new("git")
+    let ls_remote_output = Command::new(git_executable)
         .args(["ls-remote", "--branches"])
         .arg(repo_str)
         .stdout(Stdio::null())
@@ -22,7 +23,7 @@ pub fn can_access_remote(
         Ok(output) if output.success() => Ok(true),
         Ok(_) if report_git_errors => Err(String::from("couldn't access repository")),
         Ok(_) => Ok(false),
-        Err(err) => Err(format!("error calling git ls-remote: {err}")),
+        Err(err) => Err(format!("error calling `{git_executable} ls-remote`: {err}")),
     }
 }
 
@@ -31,13 +32,14 @@ pub fn clone(
     target_dir: &Path,
     print_progress: bool,
     report_git_errors: bool,
+    git_executable: &String,
 ) -> Result<(), String> {
     let stdio = match print_progress {
         true => || Stdio::inherit(),
         false => || Stdio::null(),
     };
 
-    let clone_status = Command::new("git")
+    let clone_status = Command::new(git_executable)
         .arg("clone")
         .arg(repo_str)
         .arg(target_dir)
@@ -51,6 +53,6 @@ pub fn clone(
             true => Err(String::from("git clone returned non-zero status code")),
             false => Ok(()),
         },
-        Err(err) => Err(format!("error calling git clone: {err}")),
+        Err(err) => Err(format!("error calling `{git_executable} clone`: {err}")),
     }
 }